+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/OpModes/Tele_one_player.java
@@ -1,29 +1,26 @@
-package org.firstinspires.ftc.teamcode;
-
+package org.firstinspires.ftc.teamcode.OpModes;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-
 import com.qualcomm.robotcore.hardware.DcMotorEx;
 import com.qualcomm.robotcore.hardware.Gamepad;
-
 import com.qualcomm.robotcore.hardware.configuration.typecontainers.MotorConfigurationType;
 
-import org.firstinspires.ftc.teamcode.Components.DriveTrainV2;
+import org.firstinspires.ftc.teamcode.Components.DriveTrainRC;
 
-@TeleOp(name = "Avocado_Tele(robot centric)")
-public class sugoma extends LinearOpMode {
+@TeleOp(name="Avocado_Tele_1p(robot centric)")
+public class Tele_one_player extends LinearOpMode {
 
+    private DriveTrainRC chassisV2;
 
-    private DriveTrainV2 chassis;
-
-    ///private Intake intake;
-    ///private Arm arm;
+    ///private IntakeV2 intakeV2;
+    ///private ArmV2 armV2;
     private Gamepad lastGamepad1 = new Gamepad();
     private Gamepad currentGamepad1 = new Gamepad();
     private Gamepad lastGamepad2 = new Gamepad();
     private Gamepad currentGamepad2 = new Gamepad();
     DcMotorEx frontLeft,backLeft,frontRight,backRight;
+
 
     @Override
     public void runOpMode() throws InterruptedException {
@@ -31,24 +28,13 @@ public class sugoma extends LinearOpMode {
 
         waitForStart();
 
-        while (opModeIsActive()) {
-
-            lastGamepad1.copy(currentGamepad1);
-            lastGamepad2.copy(currentGamepad2);
-
-            chassis.goGoVrumVrumV2(lastGamepad1, currentGamepad1);
-            ///arm.armControl(currentGamepad2);
-            ///intake.aspirator(currentGamepad2);
-
-            currentGamepad1.copy(gamepad1);
-            currentGamepad2.copy(gamepad2);
-
-            telemetry.update();
+        while (opModeIsActive()){
+            chassisV2.goGoVrumVrumRC(currentGamepad1,lastGamepad1);
 
         }
-    }
 
-    private void initializeHardware() {
+        }
+    public void initializeHardware(){
         telemetry.addData("0::---------------------------:",0);
         telemetry.addData("Gamepad1 input",currentGamepad1);
         telemetry.addData("Gaempad2 input",currentGamepad2);
@@ -58,6 +44,7 @@ public class sugoma extends LinearOpMode {
         frontRight = hardwareMap.get(DcMotorEx.class,"frontRight");
         backLeft = hardwareMap.get(DcMotorEx.class,"backLeft");
         backRight = hardwareMap.get(DcMotorEx.class,"backRight");
+
 
 
         MotorConfigurationType m= frontLeft.getMotorType();
@@ -75,11 +62,9 @@ public class sugoma extends LinearOpMode {
         ///rotateServo = hardwareMap.get(Servo.class,"rotateServo");
         ///intakeServo = hardwareMap.get(Crservo.class,"intakeServo");
 
-        chassis = new DriveTrainV2(frontLeft, frontRight, backLeft, backRight);
-        ///intake = new Intake(intakeServo,rotateServo);
-        ///arm = new Arm(extendMotor, slideShiftMotor);
-
-
-
+        chassisV2 = new DriveTrainRC(frontLeft, frontRight, backLeft, backRight);
+        ///intakeV2 = new IntakeV2(intakeServo,rotateServo);
+        ///armV2 = new ArmV2(extendMotor, slideShiftMotor);
     }
-}
+    }
+
